+++ apps/webapp/app/routes/orgs.$organizationSlug.usage.ts
@@ -1,0 +1,7 @@
+import { redirect, type LoaderFunctionArgs } from "@remix-run/server-runtime";
+import { OrganizationParamsSchema, v3UsagePath } from "~/utils/pathBuilder";
+
+export const loader = async ({ params }: LoaderFunctionArgs) => {
+  const { organizationSlug } = OrganizationParamsSchema.parse(params);
+  return redirect(v3UsagePath({ slug: organizationSlug }));
+};
